lttng add-trigger: parse condition / action name as option arguments
Change the `add-trigger` command's --condition and --action options to
take the condition name and action name as option arguments.  Currently,
when we see the --condition option, we go into the parse_condition
function and expect the first argument to be the condition name (same
idea for the action).  Since the condition/action name is mandatory
after --condition/--action, it is simpler to say that --condition and
--action require an argument.

This will let the user do:

    --condition=on-event

in addition to the current syntax:

    --condition on-event

The code is a bit cleaner as a result.

Change-Id: Ic7ea97c2657e5de6ae18563694ad22bbffbb2aa8
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>