Fix: sessiond: trigger with condition not requiring event notifier cannot be listed
Observed issue
==============

The lttng-sessiond process aborts when an lttng error query is performed
against a trigger that uses a condition other than "event-match".

Cause
=====

The trigger's condition is a LTTNG_CONDITION_TYPE_SESSION_CONSUMED_SIZE
which does not have a "domain type".

This results in the call to `abort()` in
`event_notifier_error_accounting_get_count` of the default case.

Solution
========

Check if the trigger "needs" or at least "depends" on tracer notifier.

Known drawbacks
=========

None

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I3d5f09617f95315060a611b464d1df95cb80c5bd