Fix: build: libcommon fd-tracker dependency is not available
Observed issue
==============

A build configured with:

  ./configure -disable-bin-lttng --disable-bin-lttng-crash --disable-bin-lttng-sessiond --disable-bin-lttng-relayd

Fails at build time with:

  make[3]: *** No rule to make target '../../src/common/fd-tracker/libfd-tracker.la', needed by 'libcommon.la'.  Stop.
  make[3]: *** Waiting for unfinished jobs....
  CC       lttng-elf.lo

Cause
=====

fd-tracker is required by libcommon. This is introduced by commit
8bb66c3cd60938352927ee865759433387324250 [1]

Build of libfd-tracker is disabled at the configure level by
build_lib_fd_tracker which in turn is enabled/disabled by the
--enable/disable-bin-* options.

For the observed issue, the --enable-bin-lttng-consumerd alone does not
enable the build of libfd-tracker.

Solution
========

All dependencies for libcommon are now always built. All bins require
libcommon to be present anyway.

This patch also fix a problem where the examples under the doc are build
even if liblttng-ctl is not built.

Known drawbacks
=========

None.

References
==========

[1]
http://git.lttng.org/?p=lttng-tools.git;a=commit;h=8bb66c3cd60938352927ee865759433387324250

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I94f5d7cdadcb4f8ff9c2617a675659c1f9eb4709