Add condition-targeting error query
Notifications discarded by the tracers are reported at the level of a
trigger. As those errors are specific to triggers with an "event-rule
matches" condition, they should be reported through a condition-specific
error query.

Note that a condition error query is created from a trigger: there is no
ambiguity since, unlike actions, conditions cannot be nested.

Given the proximity of the final 2.13 release, the code which populated
trigger error query results is simply used to populate the condition
error query results when the condition is of type "event-rule matches".

No trigger-scope errors can be reported for the moment. However, such
error reports will be added in the future.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Ie1ac3668142041beb6fd61574ccef506707c55b2