doc/man: use double quotes when referring to internal section
This patch adds double quotes to all the manual page internal section
references using their full name. Those references often have the
following AsciiDoc form:

    See the <<id,Full section name>> section below.

With this patch, this would be converted to:

    See the ``<<id,Full section name>>'' section below.

In the rendered manual page, before this patch:

    See the Full section name section below.
            ¯¯¯¯ ¯¯¯¯¯¯¯ ¯¯¯¯
With this patch:

    See the “Full section name” section below.

The purpose of this patch is, thanks to the change in
`doc/man/manpage.xsl`, to remove the italic style for the text of
internal links. Because there's no way to create dynamic internal links
in a manual page, this style causes internal links to look weird when
they're not a full section name, for example:

    Note that the trigger doesn't need to [...]
                  ¯¯¯¯¯¯¯
The HTML rendering of LTTng-tools manual pages can still benefit from
internal links. This patch makes it possible to add more internal links
without degrading the visual style of manual pages when rendered in a
terminal.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I1a5ef7eab7ff1e66c137e16b51a9c9074e43f583