Fix: sessiond: notification: missing unlock on client skip
Skipping a client must be performed by using the dedicated "skip_client"
label which will unlock the client's lock before continuing the loop
rather than using 'continue' directly.

Currently, a client will remain locked when an hidden trigger emits
a notification to which it is subscribed.

1458230 Missing unlock

May result in deadlock if there is another attempt to acquire the lock.

In notification_client_list_send_evaluation: Missing a release of a lock
on a path (CWE-667)

Reported-by: Coverity Scan
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I8b69395b91b0ea59ae5e0beadebd9099db623121