Tests fix: unix socket: leaked socket of connection to child
The child_connection socket is only used by the parent in the
credentials passing test. The teardown assumes the reverse which causes
the socket to be leaked.

1458471 Resource leak

The system resource will not be reclaimed and reused, reducing the
future availability of the resource.

In test_creds_passing: Leak of memory or pointers to system
resources (CWE-404)

Reported-by: Coverity Scan
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I2ead9abbfc189ffbdd71a27f6376d0b001cdc2a3